separate into tool/explore make it its own tool
