+++ rust/crates/sift_mcp/src/tool/explore/test.rs
