+++ rust/crates/sift_mcp/src/tool/mod.rs
@@ -1,2 +1,3 @@
 pub mod data;
+pub mod explore;
 pub mod list;
